client to be backed up
I don't recommend this for use. Passwords are saved in the script itself and it is not very efficient. Use something like rsnapshot/rsync instead unless for some reason you can't. If you want to avoid saving passwords in the script itself you can store the db credentials in ~/.my.cnf.