Hack for backing up a server.
Not recommended; inefficient and stores db passwords in script. Use rsnapshot or rsync instead. You will also need to create a user on the remote host (the one storing the backups) named backupper so this script can run; else change the script and use a different account.

If you need to use it, set this to run as a cronjob daily.